extract temporal mfp planning from operator
