+++ src/dataflow/src/render/upsert.rs
@@ -120,7 +120,7 @@ where
     }
     let temporal_plan = if !temporal.is_empty() {
         let temporal_mfp = expr::MapFilterProject::new(source_arity).filter(temporal);
-        Some(crate::MfpPlan::create_from(temporal_mfp).unwrap_or_else(|e| panic!("{}", e)))
+        Some(temporal_mfp.into_plan().unwrap_or_else(|e| panic!("{}", e)))
     } else {
         None
     };
@@ -300,10 +300,14 @@ where
             move |(row, time, diff)| {
                 let arena = repr::RowArena::new();
                 let mut datums_local = datums.borrow_with(&row);
-                let time_diffs = plan.evaluate(&mut datums_local, &arena, time, diff);
+                let times_diffs = plan.evaluate(&mut datums_local, &arena, time, diff);
                 // Explicitly drop `datums_local` to release the borrow.
                 drop(datums_local);
-                time_diffs.map(move |time_diff| time_diff.map(|(t, d)| (row.clone(), t, d)))
+                times_diffs.map(move |time_diff| {
+                    time_diff
+                        .map(|(t, d)| (row.clone(), t, d))
+                        .map_err(|(e, t, d)| (DataflowError::from(e), t, d))
+                })
             }
         });
